サイドバーのボタンをOuterFrameから分離した。
